+++ DebugMenu/Runtime/Elements/DebugIntArray.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using Containers;
 
 namespace DebugMenu
 {
@@ -41,7 +40,7 @@ namespace DebugMenu
         public override bool IsSaveable => false;
 
         /// <summary>現在のリスト長に合う子行を返す。</summary>
-        public override FastList<DebugElement> Children
+        public override IReadOnlyList<DebugElement> Children
         {
             get
             {